Revert "Merge pull request #636 from shatterednirvana/persistence-finall...

+++ AppDB/cassandra/cassandra_helper.rb
@@ -3,18 +3,8 @@ require 'djinn'
 require 'djinn_job_data'
 require 'helperfunctions'
 
-
-# The path on the local filesystem where the Cassandra executable can be found.
-CASSANDRA_BIN = "#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra"
-
-
-# The path on the local filesystem where the nodetool executable can be found.
-NODETOOL_BIN = "#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/nodetool"
-
-
-# The file that we should write the process id running Cassandra to.
-CASSANDRA_PID_FILE = "/var/appscale/appscale-cassandra.pid"
-
+# Whether to remove old data from a previous start
+DROP_TABLES = true
 
 def get_uaserver_ip()
   Djinn.get_nearest_db_ip
@@ -35,15 +25,15 @@ def get_local_token(master_ip, slave_ips)
     return 0
   end
     
-  slave_ips.each_with_index { |ip, index|
+  for ii in 0..slave_ips.length
     # Based on local ip return the correct token
     # This token generation was taken from:
     # http://www.datastax.com/docs/0.8/install/cluster_init#cluster-init
-    if ip == HelperFunctions.local_ip
+    if slave_ips[ii] == HelperFunctions.local_ip
       # Add one to offset the master
-      return (index + 1) * (2**127)/(1 + slave_ips.length)
+      return (ii + 1)*(2**127)/(1 + slave_ips.length)
     end
-  }
+  end
 end
 
 def setup_db_config_files(master_ip, slave_ips, creds)
@@ -54,7 +44,7 @@ def setup_db_config_files(master_ip, slave_ips, creds)
   local_token = get_local_token(master_ip, slave_ips)
 
   files_to_config = `ls #{source_dir}`.split
-  files_to_config.each { |filename|
+  files_to_config.each{ |filename|
     full_path_to_read = File.join(source_dir, filename)
     full_path_to_write = File.join(dest_dir, filename)
     File.open(full_path_to_read) { |source_file|
@@ -76,13 +66,12 @@ def start_db_master()
   Djinn.log_info("Starting up Cassandra as master")
 
   Djinn.log_run("pkill ThriftBroker")
-  if @creds["clear_datastore"]
+  if DROP_TABLES
     Djinn.log_run("rm -rf /var/appscale/cassandra*")
-    Djinn.log_run("rm -rf /opt/appscale/cassandra*")
+    Djinn.log_run("rm /var/log/appscale/cassandra/system.log")
   end
   
-  Djinn.log_run("rm /var/log/appscale/cassandra/system.log")
-  Djinn.log_run("#{CASSANDRA_BIN} start -p #{CASSANDRA_PID_FILE}")
+  Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra start -p /var/appscale/appscale-cassandra.pid")
   HelperFunctions.sleep_until_port_is_open(HelperFunctions.local_ip, 9160)
 end
 
@@ -92,21 +81,25 @@ def start_db_slave()
 
   HelperFunctions.sleep_until_port_is_open(Djinn.get_db_master_ip, 9160)
   sleep(5)
-  if @creds["clear_datastore"]
+  if DROP_TABLES
     Djinn.log_run("rm -rf /var/appscale/cassandra*")
     Djinn.log_run("rm /var/log/appscale/cassandra/system.log")
   end
-  Djinn.log_run("#{CASSANDRA_BIN} start -p #{CASSANDRA_PID_FILE}")
-  Djinn.log_run("#{CASSANDRA_BIN} start -p #{CASSANDRA_PID_FILE}")
+  Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra start -p /var/appscale/appscale-cassandra.pid")
+  Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra start -p /var/appscale/appscale-cassandra.pid")
   HelperFunctions.sleep_until_port_is_open(HelperFunctions.local_ip, 9160)
 end
 
 def stop_db_master
   Djinn.log_info("Stopping Cassandra master")
-  Djinn.log_run("cat #{CASSANDRA_PID_FILE} | xargs kill -9")
+  Djinn.log_run("cat /var/appscale/appscale-cassandra.pid | xargs kill -9")
 end
 
 def stop_db_slave
   Djinn.log_info("Stopping Cassandra slave")
-  Djinn.log_run("cat #{CASSANDRA_PID_FILE} | xargs kill -9")
+  if DROP_TABLES
+    Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/nodetool decommission -h #{HelperFunctions.local_ip} -p 6666")
+  end
+  Djinn.log_run("cat /var/appscale/appscale-cassandra.pid | xargs kill -9")
 end
+
